break cycle by removing feature_extraction -> anomaly deps, replaced by feature_extraction -> tree instead
