+++ river/anomaly/hst.py
@@ -4,56 +4,10 @@ import collections
 import functools
 import random
 
-from river import anomaly, base, tree
+from river import anomaly, base
+from river.tree.padded import PaddedBranch, make_padded_tree
 
 __all__ = ["HalfSpaceTrees"]
-
-
-class HSTBranch(tree.base.Branch):
-    def __init__(self, left, right, feature, threshold, l_mass, r_mass):
-        super().__init__(left, right)
-        self.feature = feature
-        self.threshold = threshold
-        self.l_mass = l_mass
-        self.r_mass = r_mass
-
-    @property
-    def left(self):
-        return self.children[0]
-
-    @property
-    def right(self):
-        return self.children[1]
-
-    def next(self, x):
-        """
-
-        We want to handle the case where a split feature is missing. In that case, we go down the
-        child that has been the most visited in the past.
-
-        """
-        left, right = self.children
-        try:
-            value = x[self.feature]
-        except KeyError:
-            if left.l_mass < right.l_mass:
-                return right
-            return left
-        if value < self.threshold:
-            return left
-        return right
-
-    def most_common_path(self):
-        raise NotImplementedError
-
-    @property
-    def repr_split(self):
-        return f"{self.feature} < {self.threshold:.5f}"
-
-
-class HSTLeaf(tree.base.Leaf):
-    def __repr__(self):
-        return str(self.r_mass)
 
 
 def _walk_learn(root, x, height):
@@ -93,41 +47,6 @@ def _walk_score(root, x, height, size_limit):
         depth_pow <<= 1
     score += node.r_mass * depth_pow
     return score
-
-
-def make_padded_tree(limits, height, padding, rng=random, **node_params):
-    if height == 0:
-        return HSTLeaf(**node_params)
-
-    # Randomly pick a feature
-    # We weight each feature by the gap between each feature's limits
-    on = rng.choices(
-        population=list(limits.keys()),
-        weights=[limits[i][1] - limits[i][0] for i in limits],
-    )[0]
-
-    # Pick a split point; use padding to avoid too narrow a split
-    a = limits[on][0]
-    b = limits[on][1]
-    at = rng.uniform(a + padding * (b - a), b - padding * (b - a))
-
-    # Build the left node
-    tmp = limits[on]
-    limits[on] = (tmp[0], at)
-    left = make_padded_tree(
-        limits=limits, height=height - 1, padding=padding, rng=rng, **node_params
-    )
-    limits[on] = tmp
-
-    # Build the right node
-    tmp = limits[on]
-    limits[on] = (at, tmp[1])
-    right = make_padded_tree(
-        limits=limits, height=height - 1, padding=padding, rng=rng, **node_params
-    )
-    limits[on] = tmp
-
-    return HSTBranch(left=left, right=right, feature=on, threshold=at, **node_params)
 
 
 class HalfSpaceTrees(anomaly.base.AnomalyDetector):
@@ -258,7 +177,7 @@ class HalfSpaceTrees(anomaly.base.AnomalyDetector):
         self.seed = seed
         self.rng = random.Random(seed)
 
-        self.trees: list[HSTBranch] = []
+        self.trees: list[PaddedBranch] = []
         self._tree_nodes: list[list] = []
         self.counter = 0
         self._first_window = True
